backend/requirements.txt
# Production dependencies for Render deployment
# Used via render.yaml buildCommand: pip install -r backend/requirements.txt
#
# These are ONLY what the Flask backend needs to run
# Local dev uses: ../requirements.txt (includes matplotlib, jupyter, etc.)

flask==3.0.0
flask-cors>=4.0.0
numpy>=1.20.0
scipy>=1.7.0
obspy>=1.3.0
requests>=2.26.0
boto3>=1.28.0
zstandard>=0.21.0
pytz>=2021.3
python-dotenv>=1.0.0

